renaming: homePanel -> modLauncher
